refactor(crates/server): rename `MockNotification` to `DummyNotification`
