+++ crates/server/src/manager/mod.rs
@@ -232,7 +232,7 @@ mod tests {
     use crate::{
         backend::LocalClipboardBackend,
         manager::{ClipboardManager, DEFAULT_CAPACITY},
-        notification::MockNotification,
+        notification::DummyNotification,
     };
 
     fn create_clips(n: usize) -> Vec<ClipEntry> {
@@ -241,7 +241,7 @@ mod tests {
 
     #[test]
     fn test_construction() {
-        let notification = MockNotification::default();
+        let notification = DummyNotification::default();
         let backend = Arc::new(LocalClipboardBackend::new());
         let mgr = ClipboardManager::new(backend, notification);
         assert!(mgr.is_empty());
@@ -263,7 +263,7 @@ mod tests {
     #[test]
     fn test_capacity() {
         let backend = Arc::new(LocalClipboardBackend::new());
-        let notification = MockNotification::default();
+        let notification = DummyNotification::default();
         let cap = 10;
         let mut mgr = ClipboardManager::with_capacity(backend, cap, notification);
         assert_eq!(mgr.len(), 0);
@@ -298,7 +298,7 @@ mod tests {
         let n = 20;
         let clips = create_clips(n);
         let backend = Arc::new(LocalClipboardBackend::new());
-        let notification = MockNotification::default();
+        let notification = DummyNotification::default();
         let mut mgr = ClipboardManager::new(backend, notification);
         for clip in &clips {
             let _ = mgr.insert(clip.clone());
@@ -319,7 +319,7 @@ mod tests {
         let n = 10;
         let mut clips = create_clips(n);
         let backend = Arc::new(LocalClipboardBackend::new());
-        let notification = MockNotification::default();
+        let notification = DummyNotification::default();
         let mut mgr = ClipboardManager::with_capacity(backend, 20, notification);
 
         mgr.import(&clips);
@@ -344,7 +344,7 @@ mod tests {
         let data2 = "АБВГД";
         let clip = ClipEntry::new(data1.as_bytes(), &MIME, ClipboardKind::Clipboard, None).unwrap();
         let backend = Arc::new(LocalClipboardBackend::new());
-        let notification = MockNotification::default();
+        let notification = DummyNotification::default();
         let mut mgr = ClipboardManager::new(backend, notification);
         let old_id = mgr.insert(clip);
         assert_eq!(mgr.len(), 1);
@@ -362,7 +362,7 @@ mod tests {
     #[test]
     fn test_remove() {
         let backend = Arc::new(LocalClipboardBackend::new());
-        let notification = MockNotification::default();
+        let notification = DummyNotification::default();
         let mut mgr = ClipboardManager::new(backend, notification);
         assert_eq!(mgr.len(), 0);
         assert!(!mgr.remove(43));
@@ -386,7 +386,7 @@ mod tests {
     #[test]
     fn test_clear() {
         let backend = Arc::new(LocalClipboardBackend::new());
-        let notification = MockNotification::default();
+        let notification = DummyNotification::default();
         let n = 20;
         let clips = create_clips(n);
         let mut mgr = ClipboardManager::new(backend, notification);
