+++ crates/server/src/notification/dummy.rs
