+++ crates/server/src/notification/mod.rs
@@ -1,9 +1,9 @@
 mod desktop;
-mod mock;
+mod dummy;
 mod traits;
 
 pub use self::{
     desktop::{Notification as DesktopNotification, Worker as DesktopNotificationWorker},
-    mock::Notification as MockNotification,
+    dummy::Notification as DummyNotification,
     traits::Notification,
 };
